Version 4.2.54 (squashed - based on bfe7f4af141da71aefd3af5b0733d1cb27281cf6)
Fix HConstant(double, ...) constructor (issue 3865).

Add NativeWeakMap to v8.h (Chromium issue 437416).

Performance and stability improvements on all platforms.

Cr-Commit-Position: refs/heads/candidates@{#25444}

+++ src/heap-snapshot-generator.cc
@@ -3102,7 +3102,7 @@ void HeapSnapshotJSONSerializer::SerializeString(const unsigned char* s) {
           WriteUChar(writer_, *s);
         } else {
           // Convert UTF-8 into \u UTF-16 literal.
-          unsigned length = 1, cursor = 0;
+          size_t length = 1, cursor = 0;
           for ( ; length <= 4 && *(s + length) != '\0'; ++length) { }
           unibrow::uchar c = unibrow::Utf8::CalculateValue(s, length, &cursor);
           if (c != unibrow::Utf8::kBadChar) {
